+++ processor/src/test/resources/input/com/example/render/BaseRenderComponent.java
@@ -16,6 +16,6 @@ class BaseRenderComponent
   @Override
   protected ReactNode render()
   {
-    return super.render();
+    return null;
   }
 }