+++ processor/src/test/resources/input/com/example/render/MyParent.java
@@ -7,13 +7,13 @@ import react4j.core.BaseState;
 import react4j.core.Component;
 import react4j.core.ReactNode;
 
-class MyParent
+abstract class MyParent
   extends Component<BaseProps, BaseState, BaseContext>
 {
   @Nullable
   @Override
-  protected ReactNode[] renderAsArray()
+  protected ReactNode render()
   {
-    return super.renderAsArray();
+    return null;
   }
 }